Network parser project
Plan is to divide as many functionalities into their own libraries, which can depend on eachother and be updated separately. Anything dealing with network files or simulating them should be done in C++, and anything representing that can be done in literally anything else as long as it can load a library from disk.